moved naughty checker to seperate file
